fix format string warnings (64-bit); avoid (m)syslog format string attacks

+++ ntpd/ntp_config.c
@@ -2840,8 +2840,8 @@ config_phone(
 			sys_phone[i++] = estrdup(*s);
 		else
 			msyslog(LOG_INFO,
-				"phone: Number of phone entries exceeds %d. Ignoring phone %s...",
-				COUNTOF(sys_phone) - 1, *s);
+				"phone: Number of phone entries exceeds %lu. Ignoring phone %s...",
+				(u_long)(COUNTOF(sys_phone) - 1), *s);
 		s = next_node(s);
 	}
 
@@ -2998,8 +2998,8 @@ config_ttl(
 			sys_ttl[i++] = (u_char)*curr_ttl;
 		else
 			msyslog(LOG_INFO,
-				"ttl: Number of TTL entries exceeds %d. Ignoring TTL %d...",
-				COUNTOF(sys_ttl), *curr_ttl);
+				"ttl: Number of TTL entries exceeds %lu. Ignoring TTL %d...",
+				(u_long)COUNTOF(sys_ttl), *curr_ttl);
 
 		curr_ttl = next_node(curr_ttl);
 	}
